+++ packages/animated-pose/src/types.ts
@@ -20,6 +20,7 @@ export interface AnimatedPoser extends Poser<Value, Action, AnimatedPoser> {
 export type TransitionProps = {
   value: Animated.Value;
   toValue: number;
+  [key: string]: any;
 };
 
 export type Transition = (
@@ -32,3 +33,21 @@ export type CreateValueProps = {
 };
 
 export type AnimatedPoseConfig = PoserConfig<Value>;
+
+export type AnimatedFactoryConfig = {
+  convertUnitToPoints: (v: string) => number;
+  unitConverters: UnitConverterMap;
+};
+
+export type AnimatedPoserFactory = (
+  config: AnimatedPoseConfig
+) => Poser<Value, Action, AnimatedPoser>;
+
+export type UnitConverter = (v: number) => number;
+
+export type DimensionConverterFactory = (
+  viewport: 'window' | 'screen',
+  axis: 'width' | 'height'
+) => UnitConverter;
+
+export type UnitConverterMap = { [key: string]: UnitConverter };