+++ packages/animated-pose/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Animated Pose adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.5] 2018-04-23
+
+- `useNativeDriver` override in all transitions
+
+## [0.0.4] 2018-04-23
+
+- Allowing `useNativeDriver` to override default transitions
+
 ## [0.0.1] 2018-04-18
 
 - First publish